Unity inc (#159)
* use a bit more add_inplace

* use blas even short sequences

* enable openblas for x8 speedup :-)

* add_inplace everywhere !

* cleanup deps

* more openblas

* fix test_WaveformToFbankConverter

* optimize (B, 1, D_out) @ (D_in, D_out) matmul

* tweak_lprobs

* SC fixes

* add vocab in the model

* add test_tokenizer

* detokenize

* handle leading space

* format

ruff check --fix --select I *.py
ruff format *.py

* WIP fix tests

* add tracy profiling

* sort cpp imports

* CLI & bug fix

* fmt & revert unnecessary changes

* rm remove_head_row & get_first_cols_by_rows

* cli: use fairseq2_spm_detokenize

* use more fine grain context for beam search

lower memory usage from xGB to <0.5GB

* handle several input files

---------

+++ ggml/third_party_ggml.py
@@ -774,7 +774,7 @@ class ggml_init_params(ctypes.Structure):
     """
 
     _fields_ = [
-        ("mem_size", ctypes.c_size_t),
+        ("mem_size", ctypes.c_int64),
         ("mem_buffer", ctypes.c_void_p),
         ("no_alloc", ctypes.c_bool),
     ]
@@ -1257,7 +1257,7 @@ def ggml_get_mem_size(ctx: ggml_context_p) -> int:
 
 
 lib.ggml_get_mem_size.argtypes = [ggml_context_p]
-lib.ggml_get_mem_size.restype = ctypes.c_size_t
+lib.ggml_get_mem_size.restype = ctypes.c_int64
 
 
 # GGML_API size_t  ggml_get_max_tensor_size(const struct ggml_context * ctx);
